reuse  common reducer function across state
